Update the buildtask, but delete not working
