+++ src/index.js
@@ -1,7 +1,7 @@
 import './style.css';
 import todoTasks from './modules/taskDB';
 import {
-  addTask,
+  buildTask,
   addNewTask,
   deleteTask,
   toggleEditMode,
@@ -12,25 +12,7 @@ const listDiv = document.getElementById('list');
 const clearBtn = document.getElementById('clear-completed');
 const inputField = document.getElementById('newtask');
 
-// Function to build the tasks on the screen
-const buildTask = () => {
-  listDiv.innerHTML = ''; // Clear existing tasks before building
-  todoTasks.forEach((task) => {
-    addTask(task.description, task.index, task.completed, listDiv);
-  });
-  // Get checkboxes after building the tasks
-  const checkboxes = document.querySelectorAll('.check-box');
-  // Add event listener for the 'change' event on the checkboxes
-  checkboxes.forEach((checkbox) => {
-    checkbox.addEventListener('change', () => {
-      const index = Number(checkbox.id.split('-')[0]);
-      const status = todoTasks[index - 1].completed;
-      todoTasks[index - 1].completed = !status;
-      saveToLocalStorage();
-    });
-  });
-  saveToLocalStorage();
-};
+
 
 // Add event listener for the 'keydown' event on the input field
 inputField.addEventListener('keydown', (event) => {
@@ -46,7 +28,7 @@ clearBtn.addEventListener('click', () => {
 
   if (completedIndexes.length > 0) {
     deleteTask(completedIndexes);
-    buildTask(); // Rebuild the tasks after
+    buildTask(listDiv); // Rebuild the tasks after
     // deleting completed tasks
   }
 });
@@ -67,7 +49,7 @@ listDiv.addEventListener('click', (event) => {
       toggleEditMode(taskElement, index);
     } else if (dataAction === 'delete') {
       deleteTask([index - 1]); // Use deleteTask function to remove the task
-      buildTask(); // Rebuild the tasks after deleting
+      buildTask(listDiv); // Rebuild the tasks after deleting
     }
   } else if (target.tagName === 'P') {
     toggleEditMode(taskElement, index); // Toggle to edit mode when the task description is clicked
@@ -83,5 +65,5 @@ window.addEventListener('DOMContentLoaded', () => {
       todoTasks.push(task);
     });
   }
-  buildTask();
+  buildTask(listDiv);
 });
